+++ polyA/load_alignments.py
@@ -152,8 +152,8 @@ def load_alignments(
                     strand=meta["SD"],
                     flank=int(meta["FL"]),
                     sub_matrix_name=meta["MX"],
-                    gap_init=int(meta["GI"]),
-                    gap_ext=int(meta["GE"]),
+                    gap_init=float(meta["GI"]),
+                    gap_ext=float(meta["GE"]),
                 )
             else:
                 yield Alignment(
@@ -170,8 +170,8 @@ def load_alignments(
                     strand=meta["SD"],
                     flank=int(meta["FL"]),
                     sub_matrix_name=meta["MX"],
-                    gap_init=int(meta["GI"]),
-                    gap_ext=int(meta["GE"]),
+                    gap_init=float(meta["GI"]),
+                    gap_ext=float(meta["GE"]),
                 )
 
             meta.clear()
